+++ soxspipe/commonutils/detect_continuum.py
@@ -502,6 +502,13 @@ class detect_continuum(_base_detect):
         self.axisBDeg = self.recipeSettings["disp-axis-deg"]
         self.orderDeg = self.recipeSettings["order-deg"]
 
+        home = expanduser("~")
+        self.qcDir = self.settings["workspace-root-dir"].replace("~", home) + f"/qc/{self.recipeName}/"
+        self.qcDir = self.qcDir.replace("//", "/")
+        # RECURSIVELY CREATE MISSING DIRECTORIES
+        if not os.path.exists(self.qcDir):
+            os.makedirs(self.qcDir)
+
         return None
 
     def get(self):
@@ -645,6 +652,9 @@ class detect_continuum(_base_detect):
             "reduction_date_utc": utcnow,
             "product_desc": product_desc,
             "file_path": plotPath
+            "product_desc": f"Residuals of the order centre polynomial fit",
+            "file_path": plotPath,
+            "label": "QC"
         }).to_frame().T], ignore_index=True)
 
         mean_res = np.mean(np.abs(orderPixelTable[f'cont_{self.axisA}_fit_res'].values))
@@ -1074,14 +1084,7 @@ class detect_continuum(_base_detect):
         else:
             filename = self.sofName + "_OBJECT_TRACE_residuals.pdf"
 
-        home = expanduser("~")
-        outDir = self.settings["workspace-root-dir"].replace("~", home) + "/qc/pdf"
-        outDir = outDir.replace("//", "/")
-        # Recursively create missing directories
-        if not os.path.exists(outDir):
-            os.makedirs(outDir)
-
-        filePath = f"{outDir}/{filename}"
+        filePath = f"{self.qcDir}/{filename}"
         plt.tight_layout()
         plt.savefig(filePath, dpi=720)