+++ soxspipe/commonutils/tests/test_data_organiser.py
@@ -35,7 +35,8 @@ pathToOutputDir = home + "/xshooter-pipeline-data/unittest_data/xsh/data-organis
 
 try:
     shutil.rmtree(pathToOutputDir)
-except:
+except Exception as e:
+    print(e)
     pass
 # COPY INPUT TO OUTPUT DIR
 shutil.copytree(pathToInputDir, pathToOutputDir)
@@ -49,6 +50,7 @@ class test_data_organiser(unittest.TestCase):
 
     import pytest
 
+    @pytest.mark.full
     def test_data_organiser_function(self):
 
         from soxspipe.commonutils import data_organiser