+++ soxspipe/commonutils/tests/test_horne_extraction.py
@@ -86,7 +86,8 @@ class test_horne_extraction(unittest.TestCase):
             "file_type": [],
             "obs_date_utc": [],
             "reduction_date_utc": [],
-            "file_path": []
+            "file_path": [],
+            "label": []
         })
 
         from os.path import expanduser