+++ soxspipe/recipes/soxs_mdark.py
@@ -228,7 +228,8 @@ class soxs_mdark(_base_recipe_):
             "obs_date_utc": self.dateObs,
             "reduction_date_utc": utcnow,
             "product_desc": f"{self.arm} Master dark frame",
-            "file_path": productPath
+            "file_path": productPath,
+            "label": "PROD"
         }).to_frame().T], ignore_index=True)
 
         self.report_output()